requirements.txt
flask==2.3.3
Flask_SQLAlchemy==3.0.5
Jinja2==3.1.2
python-dotenv==1.0.0
werkzeug==2.3.7
Flask-Login==0.6.2
